fix: entity/DDL alignment, BigDecimal types, PagoService refactor, tooling scripts
- Align all entity @Column annotations with DDL (precision, scale, unique, nullability)
- Fix Float/Double → BigDecimal in ms-espacios, ms-tarifas, ms-pagos entities and services
- Refactor PagoService.crear() into 10 private methods; fix integer division bug (minutos/60)
- Add start-all.ps1 and load-db.ps1 utility scripts
- Rewrite db/00_run_all.sql as a single inlined file (phpMyAdmin + CLI compatible)
- Add docs/audit-schema.html with DDL vs entity audit report (updated to reflect fixes)
- Update README with new scripts documentation

+++ ms-espacios/src/main/java/estacionamientos/ms_espacios/dto/TipoEspacioResponseDTO.java
@@ -1,5 +1,7 @@
 package estacionamientos.ms_espacios.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,5 +14,5 @@ public class TipoEspacioResponseDTO {
     private Long id;
     private String nombre;
     private String descripcion;
-    private Double factorPrecio;
+    private BigDecimal factorPrecio;
 }
